third_party/WebKit/LayoutTests/editing/execCommand/insert-list-and-strikethrough.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  [
    '<div contenteditable>',
      '<strike>^Female</strike>',
      '<dl><strike>DT element</strike></dl>',
      '<strike><br></strike>',
      '<hr>|',
    '</div>'
  ].join(''),
  'insertOrderedList',
  [
    '<div contenteditable>',
      '<ol><li><strike>^Female</strike></li></ol>',
      '<dl>',
        '<ol>',
          '<ul><strike>DT element</strike></ul>',
          '<li><strike><br></strike></li>',
          '<li><hr>|</li>',
        '</ol>',
      '</dl>',
    '</div>'
  ].join('')),
  'insertOrderedList with STRIKE, DL, and HR');
</script>
